requirements.txt
fastapi
requests
streamlit
uvicorn
python-dotenv
langserve
langchain_community
langchain_core